feat: Better extraction and health bar
WorldData was created to store things like the gravity. It is an autoloaded field
Added a PlayerStats screen which also houses a health bar for the player
Added a Healthbar which connects with PlayerData
PlayerData now keeps track of the player's health so this was removed from Actor.gd and Player.gd
Made Actor_Enemy because multi-inheritance isn't supported. Basically the same as Actor just specific to the enemies
Changed player and enemies to use new WorldData