fix: improved ladder
The Ladder.gd represents a single ladder.
Added LadderTop, LadderMiddle and LadderBottom to represent each component individually so that ladders of varying sizes can be made.
Can't use Tilemap because they don't support Area2D nodes, only StaticBody2D
Getting on and off ladders is now triggered by a signal in PlayerData

Matt taught me that changing the _on_ladder state should be handled by an into not a bool. This is because $ObjectDetector.get_overlapping_bodies() isn't updated while you are looping. Which meant the player didn't always detect that it was still on the ladder.