feat: added melee weapons
Added the weapon and weapon_path fields to the player. Currently only works for swords
Gave the required fields to the enemies as well
Added a method to determine if the player can perform an attack. Currently only works for melee weapons
The Players BasicStateMachine now changes the position of both Position2Ds and their children when turning
The sword goes invisible (and inactive) when the player is on a ladder
Added group tags to all objects and characters
Added an attack cooldown because the sword seems to get stuck if swung too frequently
The sword controls its attack direction and has its own state machine